rename schema folder, update imports
